chore: Add correct classNames in image component
