--- docs/modules/pages/components/opinions/ts/store/index.ts
@@ -1,2 +1,0 @@
-import type { IWidgetStore } from "@beyond-js/widgets/controller";
-export class StoreManager implements IWidgetStore {}
