+++ project/modules/image/module.json
@@ -1,24 +1,16 @@
 {
-  "name": "image",
-  "platforms": [
-    "web",
-    "ios",
-    "android"
-  ],
-  "code": {
-    "hmr": true,
-    "name": "image",
-    "ts": {
-      "path": "/code/ts",
-      "files": [
-        "*"
-      ]
-    },
-    "sass":{
-      "path": "/code/scss",
-      "files": [
-        "*"
-      ]
-    }
-  }
+	"name": "image",
+	"platforms": ["web", "ios", "android"],
+	"code": {
+		"hmr": true,
+		"name": "image",
+		"ts": {
+			"path": "/ts",
+			"files": ["*"]
+		},
+		"sass": {
+			"path": "/scss",
+			"files": ["*"]
+		}
+	}
 }
