+++ project/modules/image/scss/style.scss
@@ -1,4 +1,4 @@
-.beyond-element-image {
+.pui-image {
 	position: relative;
 	-webkit-margin-before: 0;
 	margin-block-start: 0;
@@ -9,8 +9,8 @@
 	-webkit-margin-end: 0;
 	margin-inline-end: 0;
 
-	&.beyond-element-image-error,
-	&.beyond-element-image-preload {
+	&.pui-image-error,
+	&.pui-image-preload {
 		background-color: var(--gray);
 		display: block;
 		img {
@@ -18,7 +18,7 @@
 			height: 100%;
 		}
 	}
-	&.beyond-element-image-error {
+	&.pui-image-error {
 		display: flex;
 		align-items: center;
 		justify-content: center;
